+++ src/components/MyReservations/ReservationCard.tsx
@@ -1,7 +1,7 @@
 import Button from "@/src/components/Button/Button";
 import PopupModal from "@/src/components/Modal/PopupModal";
 import ReviewModal from "@/src/components/Modal/ReviewModal/ReviewModal";
-import RESERVATION_LABEL from "@/src/constants/reserbationStatus";
+import RESERVATION_LABEL from "@/src/constants/reservationStatus";
 import { usePatchMyReservation } from "@/src/queries/useMyReservations";
 import useModalStore from "@/src/stores/ModalStore";
 import { Reservation } from "@/src/types/myReservationsResponses";
